Add tools subdir for cli tools, with bash-completion
This change adds a directory where all of the tools that are shared
by different operation systems should go.

Within the Makefile, there is an OS switch which sets and OS install
command and then uses the command from a variable for many different
tools. Currently, the only tool that is installed is bash-completion.