Reinstall git after installing bash-completion
Prior to this change, git would be installed initially to close the
gohdotfiles repository. This would be before bash-completion would be
installed, too, so the bash-completion scripts wouldn't be installed.

This change reinstalls git through the os package manager so that
the completion scripts are installed. It also means that, for macOS,
git is not installed and managed by Homebrew, which will make it easier
to upgrade it over time.