Create alias for hostname-bash-profile
Prior to this change, there was no quick way of referencing the hostname
profile whilst in a shell, you would have to go to the gohbashprofile
directory and find it manually.

This change creates an alias that will calculate the profile location
when it is run. It is done at runtime so that it is evaluated as the
shell would evaluate it when running the bash profile. This way, the
hostname can change and the alias will return the correct value.