Don't overwrite standard commands with flags
Prior to this commit, there were a couple of aliases in the profile that
would overwrite the standard commands with the same command but with a
flag present.

This commit removes these aliases because it could mean that the user
gets used to the commands having the flags there and then not use them
when they could be required in another environment.

I am considering it bad practise to overwrite custom behaviour of a
binary through aliases.