Default to CURDIR for referencing profile dirs
Prior to this change, there was a repeated variable across many of the
profiles' Makefiles. This was because the absolute location of the
profile directory requires being used within some of the scripts.

The location of the directory was previously hardcoded into the
Makefiles, making it hard for the directory structure to change.

This change uses the CURDIR variable provided by GNU Make which is set
to the value of the directory from with the Make rules are being
triggered from. This means that as long as the directory is the PWD or
the Makefile is triggered using the '-C' flag, like `make -C DIR` then
CURDIR will be set to the correct value.