Deduce GOH_BASHPROFILE_DIR from $GOH_BASHPROFILE
Prior to this commit, the GOH_BASHPROFILE_DIR env var was deduced from
an env var GOH_DOTFILES. This wasn't the best way to do it as the Bash
profile shouldn't be required to depend on the whole, DOTFILES
directory, only the BASHPROFILE directory.

With the new way of deducing it from the GOH_BASHPROFILE, it makes it
easier to install only the bash profile and not have to have the
DOTFILE directory as well.